Add PegasusQ Governor
PegasusQ is a multi-core governor, based on the Ondemand governor and governor with integrated hot-plugging.
Ongoing processes in the queue, we know that multiple processes can run simultaneously on. These processes are active in an array, which is a field called "Run Queue" that is ongoing, with their priority values ​​arranged (priority will be used by the task scheduler, which then decides which process to run next.

Signed-off-by: ROM-Jeremy <Disciple97414@Gmail.com>