Trying to enable call recording
Finding these patches on github was a pain in the ass. These were the same patches used in 3 different kernels I found.
This will more than likely not work, but if it does, that will be great!
These are the patches from skvalex, and from reading through xda, he has enabled call recording on quite a bit of devices.
So I used the same method he used, hoping it will work

Signed-off-by: ROM-Jeremy <Disciple97414@Gmail.com>